Fix the Ofx#accounts method
This was returning nil in most cases because it was written with `compact!` instead of `compact` - the bang variant is intended for mutating the original array rather than returning a new one, and as such it is designed to return `nil` when no changes are made. So in this case, we don't want to be using the bang variant. Likewise for `flatten!` -> `flatten`.