Stop doing any amount processing based on TRNTYPE
Karl found that the OFX specification says not to use the TRNTYPE as any indicator of the actual sign of the transaction https://www.ofx.net/downloads/OFX%202.2.pdf :

"TRNTYPE: Transaction type, see section 11.4.4.3 for possible values. This element does not change the effect of the transaction upon the balance (increases and decreases are indicated by the sign of the <TRNAMT>)"

A user hit this problem where the TRNTYPE of refunds were still DEBIT, though the amount was actually positive - meaning it was all messed up.

Suspicion is that this processed_amount stuff was added pre-emptively, or back in 2008 when banks didn't follow OFX specifications too well. This commit removes that stuff, and just passes the transaction sign through natively through .to_f.